+++ docs/changes.rst
@@ -6,6 +6,7 @@ Not yet released.
 * Added support for browsing changes for a individual string.
 * Fixed plurals handling in automatic translation from other components.
 * Added keyboard shortcut Alt+Enter to submit string as a suggestion.
+* Added support for placeables in the Fluent format.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/88?closed=1>`__.